+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `cssPropertyDistributor` utility.
+
 ## [2.60.2] - 2020-03-26
 
 Fix: